Revert "injected the correct time to the gpsOne engine"
This reverts commit c179a2a637f03e1d5f7e9870da542d8cb4d2f594.

+++ leo-gps.c
@@ -984,7 +984,7 @@ static int gps_stop() {
 
 static int gps_inject_time(GpsUtcTime time, int64_t timeReference, int uncertainty) {
     D("%s() is called", __FUNCTION__);
-    D("time=%lld, timeReference=%lld, uncertainty=%d", time, timeReference, uncertainty);
+    D("time=%d, timeReference=%d, uncertainty=%d", (int) time, (int) timeReference, uncertainty);
     GpsState*  s = _gps_state;
     if (!s->init)
         return 0;
